Create checksum and password hash methods
